fix : 삭제 시 soft delete 로 변경

+++ src/main/java/com/sparta/newsfeedproject/service/UserService.java
@@ -4,7 +4,8 @@ import com.sparta.newsfeedproject.domain.User;
 import com.sparta.newsfeedproject.dto.request.*;
 import com.sparta.newsfeedproject.config.PasswordEncoder;
 import com.sparta.newsfeedproject.dto.response.UserResponseDto;
-import com.sparta.newsfeedproject.exception.DeleteException;
+import com.sparta.newsfeedproject.exception.DeletedAccountException;
+import com.sparta.newsfeedproject.exception.DeletedException;
 import com.sparta.newsfeedproject.exception.InvalidPasswordException;
 import com.sparta.newsfeedproject.exception.NotFoundException;
 import com.sparta.newsfeedproject.repository.UserRepository;
@@ -69,19 +70,21 @@ public class UserService {
     }
 
     @Transactional
-    public void userWithdrawal(User tokenUser, UserWithdrawalRequestDto userWithdrawalRequestDto) {
-        User user = userRepository.findById(tokenUser.getId()).orElseThrow(() -> new NotFoundException("해당 유저를 찾을 수 없습니다."));
+    public void userWithdrawal(UserTokenDto tokenUser, UserWithdrawalRequestDto userWithdrawalRequestDto) {
+
+        if(!userRepository.existsByIdAndDeletedAtIsNotNull(tokenUser.getUserId())) throw new DeletedAccountException(HttpStatus.BAD_REQUEST, "이미 삭제된 계정입니다.");
+
+        User user = userRepository.findById(tokenUser.getUserId()).orElseThrow(() -> new NotFoundException("해당 유저를 찾을 수 없습니다."));
 
         try {
-
             boolean checkPassword = passwordEncoder.matches(userWithdrawalRequestDto.getPassword(), user.getPassword());
             if(checkPassword){
                 userRepository.delete(user);
             }else{
                 throw new InvalidPasswordException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
             }
-        } catch (DeleteException e) {
-            throw new DeleteException(HttpStatus.INTERNAL_SERVER_ERROR, "삭제도중 알수없는 오류가 발생하였습니다.");
+        } catch (RuntimeException e) {
+            throw new DeletedException(HttpStatus.INTERNAL_SERVER_ERROR, "삭제도중 알수없는 오류가 발생하였습니다.");
         }
     }
